Added types to increase compilation performance.

+++ src/Treatment/Treatment.jl
@@ -1,9 +1,9 @@
 #We define what a treatment is
 mutable struct Treatment
-    detecting_size::Int
-    starting_size::Int
-    pausing_size::Int 
-    resistance_gene::Int
+    detecting_size::Int64
+    starting_size::Int64
+    pausing_size::Int64
+    resistance_gene::Int64
     kill_rate::Float64
     active::Bool
     detected::Bool
